requirements.txt
Django==1.11.1
argparse==1.2.1
certifi==2017.4.17
chardet==3.0.3
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.3
gevent==1.2.1
greenlet==0.4.12
gunicorn==19.7.1
idna==2.5
pytz==2017.2
requests==2.17.3
urllib3==1.21.1
wsgiref==0.1.2
